ADD Refactor to extract common command line functionality
